Add .babelrc file to example/basic
Without the .babelrc, when trying to install and run just the example
an error occurs stating:

Couldn't find preset "es2015" relative to directory
"<DIR>/redux-sample-router"

This is due to babel walking up the directory tree looking for a
.babelrc file to use. Once it finds the projects folder it would appear
to start looking from that points up again to resolve the presets.

Being that the .babelrc file needs to be in the same directory as the
node_modules directory anyways, I've copied the file into the
example/basic folder.