Compare location changes against initialState.
Also, replace the location, rather than transitionTo it.

The location in the initialState is no longer undefined, so we should compare if we're going back to that state. Because the initial action is normally a POP, we can't transition to that kind of action and instead need to replace our location.

Fixes #164
Fixes #252